Refakturering av MigreringService, flytter valideringsmetoder til InfotrygdPeriodeMigreringService for å enklere kunne bruke den fra annet sted for å kunne validere periode
