+++ src/main/kotlin/no/nav/familie/ef/sak/behandling/migrering/MigreringService.kt
@@ -1,5 +1,6 @@
-package no.nav.familie.ef.sak.behandling
+package no.nav.familie.ef.sak.behandling.migrering
 
+import no.nav.familie.ef.sak.behandling.BehandlingService
 import no.nav.familie.ef.sak.behandling.domain.Behandling
 import no.nav.familie.ef.sak.behandling.domain.BehandlingResultat
 import no.nav.familie.ef.sak.behandling.domain.BehandlingStatus
@@ -18,7 +19,6 @@ import no.nav.familie.ef.sak.fagsak.domain.FagsakPerson
 import no.nav.familie.ef.sak.fagsak.domain.Stønadstype
 import no.nav.familie.ef.sak.fagsak.dto.MigreringInfo
 import no.nav.familie.ef.sak.infotrygd.InfotrygdService
-import no.nav.familie.ef.sak.infotrygd.InfotrygdStønadPerioderDto
 import no.nav.familie.ef.sak.infotrygd.SummertInfotrygdPeriodeDto
 import no.nav.familie.ef.sak.infrastruktur.exception.ApiFeil
 import no.nav.familie.ef.sak.infrastruktur.exception.feilHvisIkke
@@ -35,18 +35,14 @@ import no.nav.familie.ef.sak.vedtak.dto.ResultatType
 import no.nav.familie.ef.sak.vedtak.dto.VedtaksperiodeDto
 import no.nav.familie.ef.sak.vedtak.dto.tilPerioder
 import no.nav.familie.ef.sak.vilkår.VurderingService
-import no.nav.familie.kontrakter.ef.felles.StønadType
 import no.nav.familie.kontrakter.ef.infotrygd.InfotrygdEndringKode
 import no.nav.familie.kontrakter.ef.infotrygd.InfotrygdPeriode
-import no.nav.familie.kontrakter.ef.infotrygd.InfotrygdSak
-import no.nav.familie.kontrakter.ef.infotrygd.InfotrygdSakResultat
 import no.nav.familie.prosessering.domene.TaskRepository
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.math.BigDecimal
-import java.time.LocalDate
 import java.time.YearMonth
 import java.util.UUID
 
@@ -65,37 +61,16 @@ class MigreringService(
         private val featureToggleService: FeatureToggleService,
         private val infotrygdService: InfotrygdService,
         private val beregningService: BeregningService,
-        private val simuleringService: SimuleringService
+        private val simuleringService: SimuleringService,
+        private val infotrygdPeriodeMigreringService: InfotrygdPeriodeMigreringService
 ) {
 
     private val logger = LoggerFactory.getLogger(javaClass)
     private val secureLogger = LoggerFactory.getLogger("secureLogger")
 
-    private enum class MigreringExceptionType {
-        ÅPEN_SAK,
-        FLERE_IDENTER,
-        FLERE_AKTIVE_PERIODER,
-        MANGLER_AKTIVE_PERIODER,
-        HAR_ALLEREDE_BEHANDLINGER,
-        FEIL_STØNADSTYPE,
-        FLERE_IDENTER_VEDTAK,
-        ALLEREDE_MIGRERT,
-        MANGLER_PERIODER,
-        MANGLER_PERIODER_MED_BELØP,
-        FEIL_FOM_DATO,
-        FEIL_TOM_DATO,
-        ELDRE_PERIODER,
-        SIMULERING_FEILUTBETALING,
-        SIMULERING_ETTERBETALING,
-        BELØP_0,
-    }
-
-    private class MigreringException(val årsak: String, val type: MigreringExceptionType) : RuntimeException(årsak)
-
     fun hentMigreringInfo(fagsakPersonId: UUID, kjøremåned: YearMonth = kjøremåned()): MigreringInfo {
         val periode = try {
             val fagsakPerson = fagsakPersonService.hentPerson(fagsakPersonId)
-            validerSakerIInfotrygd(fagsakPerson)
             hentGjeldendePeriodeOgValiderState(fagsakPerson, kjøremåned)
         } catch (e: MigreringException) {
             logger.info("Kan ikke migrere fagsakPerson=$fagsakPersonId årsak=${e.type}")
@@ -123,7 +98,6 @@ class MigreringService(
     @Transactional
     fun migrerOvergangsstønad(fagsakPersonId: UUID): UUID {
         val fagsakPerson = fagsakPersonService.hentPerson(fagsakPersonId)
-        validerSakerIInfotrygd(fagsakPerson)
         val personIdent = fagsakPerson.hentAktivIdent()
         val kjøremåned = kjøremåned()
         val fagsak = fagsakService.hentEllerOpprettFagsak(personIdent, Stønadstype.OVERGANGSSTØNAD)
@@ -205,7 +179,7 @@ class MigreringService(
     fun erOpphørtIInfotrygd(behandlingId: UUID, opphørsmåned: YearMonth): Boolean {
         val opphørsdato = opphørsmåned.atEndOfMonth()
         val personIdent = behandlingService.hentAktivIdent(behandlingId)
-        val perioder = hentPerioder(personIdent)
+        val perioder = infotrygdService.hentDtoPerioder(personIdent).overgangsstønad
         val overførtNyLøsningOpphørsdato =
                 perioder.perioder.find { it.kode == InfotrygdEndringKode.OVERTFØRT_NY_LØSNING }?.opphørsdato
         val maxStønadTom = perioder.summert.maxOf { it.stønadTom }
@@ -236,129 +210,7 @@ class MigreringService(
         val personIdent = fagsakPerson.hentAktivIdent()
         val fagsak = fagsakService.finnFagsakerForFagsakPersonId(fagsakPerson.id).overgangsstønad
         fagsak?.let { validerFagsakOgBehandling(it) }
-        val perioder = hentPerioder(personIdent)
-        validerHarKunEnIdentPåPerioder(perioder, personIdent)
-        return periodeFremITidenEllerBakITiden(perioder, kjøremåned)
-    }
-
-    private fun periodeFremITidenEllerBakITiden(perioder: InfotrygdStønadPerioderDto,
-                                                kjøremåned: YearMonth): SummertInfotrygdPeriodeDto {
-        val gjeldendePerioder = perioder.summert
-        val perioderFremITiden = gjeldendePerioder.filter { it.stønadTom >= førsteDagenINesteMåned(kjøremåned) }
-        if (perioderFremITiden.isNotEmpty()) {
-            return gjeldendePeriodeFremITiden(perioderFremITiden, kjøremåned)
-        }
-
-        if (!featureToggleService.isEnabled("familie-ef-sak.migrering-bak-i-tiden")) {
-            throw MigreringException("Mangler aktive perioder", MigreringExceptionType.MANGLER_AKTIVE_PERIODER)
-        }
-
-        if (gjeldendePerioder.isEmpty()) {
-            throw MigreringException("Har ikke noen perioder å migrere",
-                                     MigreringExceptionType.MANGLER_PERIODER)
-        }
-        val perioderMedBeløp = gjeldendePerioder.filter { it.beløp != 0 }
-        if (perioderMedBeløp.isEmpty()) {
-            throw MigreringException("Har ikke noen perioder med beløp å migrere",
-                                     MigreringExceptionType.MANGLER_PERIODER_MED_BELØP)
-        }
-        val sistePeriode = perioderMedBeløp.maxByOrNull { it.stønadFom } ?: error("Finner ikke noen perioder")
-
-        return sisteMånedenPåPeriodeBakITiden(sistePeriode)
-    }
-
-    /**
-     * Perioder frem i tiden migreres fra neste måned
-     */
-    private fun gjeldendePeriodeFremITiden(gjeldendePerioder: List<SummertInfotrygdPeriodeDto>,
-                                           kjøremåned: YearMonth): SummertInfotrygdPeriodeDto {
-        if (gjeldendePerioder.size > 1) {
-            throw MigreringException("Har fler enn 1 (${gjeldendePerioder.size}) aktiv periode",
-                                     MigreringExceptionType.FLERE_AKTIVE_PERIODER)
-        }
-        val periode = gjeldendePerioder.single()
-        validerFomDato(periode)
-        validerTomDato(periode)
-        return periode.copy(stønadFom = maxOf(kjøremåned.atDay(1), periode.stønadFom))
-    }
-
-    /**
-     * Henter siste måneden for en periode bak i tiden
-     * Hvis det kun er 1 måned, så valideres det att fom-dato ikke er annet enn 1 i måneden
-     */
-    private fun sisteMånedenPåPeriodeBakITiden(periode: SummertInfotrygdPeriodeDto): SummertInfotrygdPeriodeDto {
-        val stønadTom = periode.stønadTom
-        val stønadFom = periode.stønadFom
-        val tomMåned = YearMonth.of(stønadTom.year, stønadTom.month)
-        val nyFomDato = tomMåned.atDay(1)
-        validerTomDato(periode)
-        if (stønadFom > nyFomDato) {
-            throw MigreringException("Startdato er annet enn første i måneden, dato=$stønadFom",
-                                     MigreringExceptionType.FEIL_FOM_DATO)
-        }
-        if (periode.beløp == 0) {
-            throw MigreringException("Beløp er 0 på siste perioden, har ikke støtte for det ennå. fom=$stønadFom",
-                                     MigreringExceptionType.BELØP_0)
-        }
-        return periode.copy(stønadFom = YearMonth.of(stønadTom.year, stønadTom.month).atDay(1))
-    }
-
-    private fun validerFomDato(periode: SummertInfotrygdPeriodeDto) {
-        if (periode.stønadFom.dayOfMonth != 1) {
-            throw MigreringException("Startdato er annet enn første i måneden, dato=${periode.stønadFom}",
-                                     MigreringExceptionType.FEIL_FOM_DATO)
-        }
-    }
-
-    private fun validerTomDato(periode: SummertInfotrygdPeriodeDto) {
-        val dato = periode.stønadTom
-        if (YearMonth.of(dato.year, dato.month).atEndOfMonth() != dato) {
-            throw MigreringException("Sluttdato er annet enn siste i måneden, dato=$dato",
-                                     MigreringExceptionType.FEIL_TOM_DATO)
-        }
-        if (dato.isBefore(LocalDate.now().minusYears(3))) {
-            throw MigreringException("Kan ikke migrere når forrige utbetaling i infotrygd er mer enn 3 år tilbake i tid, dato=$dato",
-                                     MigreringExceptionType.ELDRE_PERIODER)
-        }
-    }
-
-    private fun validerHarKunEnIdentPåPerioder(perioder: InfotrygdStønadPerioderDto,
-                                               personIdent: String) {
-        perioder.perioder.find { it.personIdent != personIdent }?.let {
-            throw MigreringException("Det finnes perioder som inneholder annet fnr=${it.personIdent}. " +
-                                     "Disse vedtaken må endres til aktivt fnr i Infotrygd",
-                                     MigreringExceptionType.FLERE_IDENTER_VEDTAK)
-        }
-    }
-
-    private fun validerSakerIInfotrygd(fagsakPerson: FagsakPerson): List<InfotrygdSak> {
-        val personIdent = fagsakPerson.hentAktivIdent()
-        val sakerForOvergangsstønad =
-                infotrygdService.hentSaker(personIdent).saker.filter { it.stønadType == StønadType.OVERGANGSSTØNAD }
-        validerFinnesIkkeÅpenSak(sakerForOvergangsstønad)
-        validerSakerInneholderKunEnIdent(sakerForOvergangsstønad, personIdent)
-        return sakerForOvergangsstønad
-    }
-
-    private fun validerSakerInneholderKunEnIdent(sakerForOvergangsstønad: List<InfotrygdSak>,
-                                                 personIdent: String) {
-        sakerForOvergangsstønad.find { it.personIdent != personIdent }?.let {
-            throw MigreringException("Finnes sak med annen personIdent for personen. ${lagSakFeilinfo(it)} " +
-                                     "personIdent=${it.personIdent}. " +
-                                     "Disse sakene må oppdateres med aktivt fnr i Infotrygd",
-                                     MigreringExceptionType.FLERE_IDENTER)
-        }
-    }
-
-    private fun validerFinnesIkkeÅpenSak(sakerForOvergangsstønad: List<InfotrygdSak>) {
-        sakerForOvergangsstønad.find { it.resultat == InfotrygdSakResultat.ÅPEN_SAK }?.let {
-            throw MigreringException("Har åpen sak. ${lagSakFeilinfo(it)}", MigreringExceptionType.ÅPEN_SAK)
-        }
-    }
-
-    private fun lagSakFeilinfo(sak: InfotrygdSak): String {
-        return "saksblokk=${sak.saksblokk} saksnr=${sak.saksnr} " +
-               "registrertDato=${sak.registrertDato} mottattDato=${sak.mottattDato}"
+        return infotrygdPeriodeMigreringService.hentPeriodeForMigrering(personIdent, kjøremåned)
     }
 
     private fun validerFagsakOgBehandling(fagsak: Fagsak) {
@@ -372,13 +224,7 @@ class MigreringService(
         }
     }
 
-    private fun hentPerioder(personIdent: String): InfotrygdStønadPerioderDto {
-        val allePerioder = infotrygdService.hentDtoPerioder(personIdent)
-        return allePerioder.overgangsstønad
-    }
-
     private fun kjøremåned() = YearMonth.now()
-    private fun førsteDagenINesteMåned(yearMonth: YearMonth) = yearMonth.plusMonths(1).atDay(1)
 
     private fun til(periode: SummertInfotrygdPeriodeDto): YearMonth =
             YearMonth.of(periode.stønadTom.year, periode.stønadTom.month)
